+++ src/covertile/cycle2/tiles/carousel.py
@@ -106,7 +106,7 @@ class ICarouselTile(IPersistentCoverTile):
 
     overlay = schema.SourceText(
         title=_(u'Overlay Template'),
-        description=_(u'A Mustache-style template string, in which you can use {{title}}, {{desc}} or {{date}}'),
+        description=_(u'A Mustache-style template string, in which you can use {{title}}, {{desc}}, {{url}} or {{date}}'),
         required=False,
         default=DEFAULT_OVERLAY_TEMPLATE,
     )