+++ CHANGES.rst
@@ -6,8 +6,8 @@ There's a frood who really knows where his towel is.
 1.0a2 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Add Dutch translations.
+  [fredvd]
 
 1.0a1 (2015-05-27)
 ------------------