+++ CHANGES.rst
@@ -6,6 +6,9 @@ There's a frood who really knows where his towel is.
 1.0a2 (unreleased)
 ------------------
 
+- Register collective.js.cycle2 resources used in this package in the JavaScript registry.
+  [djowett]
+
 - Add Dutch translations.
   [fredvd]