+++ CHANGES.rst
@@ -6,14 +6,16 @@ There's a frood who really knows where his towel is.
 1.0a2 (unreleased)
 ------------------
 
-- Fix icon path after release of collective.cover b1.
-  [rodfersou]
+- Make package compatible with collective.cover >= 1.0b1.
+  It is recommended that you upgrade to the latest version of the package.
+  [rodfersou, hvelarde, fredvd]
 
 - Register collective.js.cycle2 resources used in this package in the JavaScript registry.
   [djowett]
 
 - Add Dutch translations.
   [fredvd]
+
 
 1.0a1 (2015-05-27)
 ------------------