+++ CHANGES.rst
@@ -9,6 +9,9 @@ There's a frood who really knows where his towel is.
 - Make {{url}} available in overlay templates so you can put a link to a   carousel item. 
   [fredvd]
 
+- Register collective.js.cycle2 resources used in this package in the JavaScript registry.
+  [djowett]
+
 - Add Dutch translations.
   [fredvd]